Added Serializer Class

+++ src/NameServer/ResolverInterface.java
@@ -15,7 +15,7 @@ public interface ResolverInterface extends Remote
 	 * @return
 	 * @throws RemoteException
 	 */
-	public String lookup(int nodeId) throws RemoteException;
+	public String lookup(short nodeId) throws RemoteException;
 
 	/**
 	 * For a given filename it returns the IP address from the owner of the file
@@ -26,5 +26,5 @@ public interface ResolverInterface extends Remote
 	 */
 	public String lookup(String filename) throws RemoteException, InvalidParameterException;
 
-	public void addToTree(int ID, String IP) throws RemoteException;   // todo: Remove when RMI testing is complete
+	public void addToTree(short ID, String IP) throws RemoteException;   // todo: Remove when RMI testing is complete
 }
